requirements.txt
alembic==1.15.2
aniso8601==10.0.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
cachelib==0.13.0
click==8.1.8
dotenv==0.9.9
flasgger==0.9.7.1
Flask==3.1.0
Flask-Caching==2.3.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
flask-restx==1.3.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
importlib_resources==6.5.2
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Mako==1.3.9
MarkupSafe==3.0.2
marshmallow==3.26.1
mistune==3.1.3
packaging==24.2
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
rpds-py==0.24.0
six==1.17.0
SQLAlchemy==2.0.40
typing_extensions==4.13.0
Werkzeug==3.1.3
